==> analyses/Braun_analyses/0_download_data.ipynb <==
# Set-up

import h5py
from scipy.sparse import csr_matrix
from scipy.io import mmwrite
import numpy as np
import pandas as pd
import anndata as ad
import scanpy as sc
from tqdm.auto import tqdm

indir = '/projects_ng/SC_DATA/LinnarsonBraun/'
outdir = './data/'

data_url = 'https://storage.googleapis.com/linnarsson-lab-human/HumanFetalBrainPool.h5'
data_fname = f"{indir}HumanFetalBrainPool.h5"

!mkdir -p $(dirname {data_fname})
!wget -nc -O {data_fname} {data_url}

data_url = 'https://storage.googleapis.com/linnarsson-lab-human/human_dev.h5ad'
data_fname = f"{indir}human_dev.h5ad"


!mkdir -p $(dirname {data_fname})
!wget -nc -O {data_fname} {data_url}

# **Part 1: Load the data** `.h5 -> .csv + .mtx`

filename =  f"{indir}HumanFetalBrainPool.h5"
f = h5py.File(filename, "r")


atts = pd.DataFrame([{k: f['shoji'][k].shape for k in f['shoji'].keys()}]).T.reset_index().rename({'index': 'key', 0: 'dims'}, axis=1)
atts = atts.sort_values('dims')

dim_names = {
    '51': 'annotation',
    '617': 'clusters',
    '59480': 'genes',
    '1665937': 'ncells'
}
a = atts['dims'].astype(str)
for s,r in dim_names.items():
    a = a.str.replace(s,r)
atts['dim_names'] = a
att_groups = {k: v.values for k, v in atts.groupby('dim_names')['key']}


# factors and embeddings:
dfs = dict(
    Embedding = f['shoji'][att_groups['(ncells, 2)'][0]][()],
    Factors = f['shoji'][att_groups['(ncells, 50)'][0]][()],
    Loadings = f['shoji'][att_groups['(genes, 50)'][0]][()])

# Clusters x annotations
dfs.update(dict(ClusterAnnotationPosterior = f['shoji'][att_groups['(clusters, annotation)'][0]][()]))

# Clusters x genes
dfs.update({f'ClusterGenes.{k}': f['shoji'][k][()] for k in att_groups['(clusters, genes)']})

# AnnotationInfo
AnnotationInfo = {k: f['shoji'][k][()] for k in att_groups['(annotation,)']}
dfs.update(dict(AnnotationInfo = pd.DataFrame(AnnotationInfo).apply(lambda x: x.str.decode("utf-8") , 0)))

# CellInfo
CellInfo = {k: f['shoji'][k][()] for k in att_groups['(ncells,)']}
dfs.update(dict(CellInfo = pd.DataFrame(CellInfo).apply(lambda x: x if not x.dtype == 'O' else x.str.decode("utf-8") , 0)))

# ClusterInfo
ClusterInfo = {k: f['shoji'][k][()] for k in att_groups['(clusters,)']}
dfs.update(dict(ClusterInfo = pd.DataFrame(ClusterInfo).apply(lambda x: x if not x.dtype == 'O' else x.str.decode("utf-8") , 0)))

# GeneInfo
GeneInfo = {k: f['shoji'][k][()] for k in att_groups['(genes,)']}
dfs.update(dict(GeneInfo = pd.DataFrame(GeneInfo).apply(lambda x: x if not x.dtype == 'O' else x.str.decode("utf-8") , 0)))

expression_is_loaded = False

# Expression
Expression = csr_matrix(f['shoji'][att_groups['(ncells, genes)'][0]][()])
expression_is_loaded = True

f.close()

## Save the dataframes

for k, df in tqdm(dfs.items()):
    if isinstance(df, pd.DataFrame):
        df.to_csv(f'{outdir}{k}.csv')
    else:
        np.savetxt(f'{outdir}{k}.csv', df, delimiter=',')

## Save the expression matrix

mmwrite(f'{outdir}Expression_ValidGenes.sparseMatrix.mtx', 
        Expression[:,dfs['GeneInfo']['ValidGenes']], 
        comment = 'Expression_ValidGenes')

mmwrite(f'{outdir}Expression.sparseMatrix.mtx', 
        Expression, 
        comment = 'Expression')

# **Part 2: Update the annotation**  
## `Run after processing the lineage annotation` 

cellmeta = dfs['CellInfo'].copy()

new_anno = pd.read_excel('./cluster_lineage_annotation.summary.xlsx')

cellmeta = cellmeta.merge(new_anno[['ClusterID.(PoolClean)', 'LongName', 'ShortName', 'discarded']], 
                          left_on='Clusters', right_on = 'ClusterID.(PoolClean)', sort=False)

cellmeta.to_csv(f'{outdir}CellInfo.annotated.csv')

# **Part 3: Save the subsets for analysis**  

import time


cellmeta.Age.nunique() == cellmeta.Age.round(2).nunique()
cellmeta.Age = cellmeta.Age.round(2).astype(str)

cellmeta.Subregion = cellmeta.Subregion.str.replace(' ', '_')

# we want to analyse the data by subregion and timepoint, and then measure celltype differences.
# celltypes that don't have enough cells are not included
# we also save the cell metadata of the subset to keep things clear
# And we don't include the discarded cells
# And we want to include only the selected regions 

selected_regions = ['Brain', 'Head', 'Forebrain', 'Telencephalon']
group_vars = ['Subregion', 'Age']
celltype_var = ['ShortName']
enough_cells = 100

meta = cellmeta.copy()
meta.discarded = meta.discarded | ~meta.Region.isin(selected_regions)

pbar = tqdm(total = (~meta[group_vars].duplicated()).sum())

for region, rmeta in meta.groupby(group_vars[0]):

    for age, ameta in rmeta.groupby(group_vars[1]):

        t0 = time.time()

        if not ameta.discarded.all():
                                    
            cinfo = ameta[celltype_var].value_counts().reset_index()
            cinfo['include'] = cinfo.iloc[:,-1] >= enough_cells
        
            included = cinfo.iloc[:,0][cinfo['include'] & ~cinfo.iloc[:,0].isna()].values
    
            selmeta = pd.concat([
                ~meta.discarded,
                meta[celltype_var].isin(included),
                meta[group_vars[0]] == region,
                meta[group_vars[1]] == age], axis=1).all(1)
    
            print(f'[{region}][{age}]')
    
            if (selmeta.sum() > 0):
    
                path = f'{outdir}/by_{group_vars[0]}/{region}/by_{group_vars[1]}/{age}/'
                !mkdir -p {path}
                
                print(f'''\
    - Starting {ameta.shape[0]} cells
    - Selected {selmeta.sum()} cells
    - Included cell types: {", ".join(included)}
    - Excluded cell types: {", ".join(cinfo.iloc[:,0][~cinfo.iloc[:,0].isin(included)])}
    -> Saving to {path}...
    ''')
    
                print('...Saving metadata')
                meta[selmeta].to_csv(f'{path}cell_meta.csv')

                
                print('...Getting expression')
                if not expression_is_loaded:
                    exp = csr_matrix(f['shoji'][att_groups['(ncells, genes)'][0]][selmeta.to_numpy(),:])
                else:
                    exp = Expression[selmeta,:]

                print('...Saving expression')
                mmwrite(f'{path}Expression.sparseMatrix.mtx', 
                        exp, 
                        comment = 'Expression')
                
                print(f'-> Done in {round(time.time()-t0,2)} seconds!#########################\n')
            else:
                print('-> Not saving\n')
    
            
        pbar.update(1)



print('Done')
