==> analyses/Braun_analyses/1_neuronal_lineage_annotation.ipynb <==
library(dplyr)

# Part 1: create an Excel with the annotation 

# Load main dataset (Table S2)
path_to_annotation_info <- '~/ewce/AnnotationInfo.csv' # <-- PLEASE UPDATE THIS PATH
# st2: Dataframe holding the initial cluster information from Table S2.
st2 <- openxlsx::read.xlsx('./science.adf1226_table_s2.xlsx')
# remove not valid clusters
st2 <- subset(st2, `ClusterID.(PoolClean)` != '--') %>%
    mutate(`ClusterID.(PoolClean)` = as.numeric(`ClusterID.(PoolClean)`) ) %>% arrange(`ClusterID.(PoolClean)`)

# Load external cluster annotation information
AnnoInfo <- read.csv(path_to_annotation_info, row.names = 1)

AnnoInfo

# Classify clusters by gene markers (Excitatory/Inhibitory)
lingenes <- list(
    'Excitatory' = c(
        'VGLUT', 'EMX1'
    ),
    'Inhibitory' = c(
        'GAD2', 'DLX'
    )
)

check_cols <- c('PoolEnriched', 'TopLevelEnriched', 'SubsetEnriched')

lin.anno.genes <- st2[,c('ClusterID.(PoolClean)', check_cols)]

paste.coma <- function(...) {paste(..., sep = ',')}

for (lin in names(lingenes)){

    for (col in check_cols){
    
        col.name <- paste(col, lin, sep = '.')
        lin.list <- list()
        
    
        for (g in lingenes[[lin]]){
            
            lin.list[[g]] <- ifelse(grepl(pattern = g, x = lin.anno.genes[,col], fixed = F), g, '')            
            
        }
        

        lin.anno.genes[, col.name] <- do.call(lin.list, what = 'paste.coma') %>%
            gsub(pattern = ',$', fixed=F, replacement = '')%>% 
            gsub(pattern = '^,', fixed=F, replacement = '')
        
        lin.anno.genes[lin.anno.genes[, col.name] == '', col.name] <- NA
        lin.anno.genes[!is.na(lin.anno.genes[, col.name]), col.name] <- paste(
            col, 
            lin.anno.genes[!is.na(lin.anno.genes[, col.name]), col.name],
            sep = ':')
        
    }
}

lin.anno.genes$gene.anno.exc <- apply(lin.anno.genes[,grepl(colnames(lin.anno.genes), pattern = 'Excitatory')], 
                                  MARGIN = 1, FUN = function(x) {
    if (all(is.na(x))){NA}else{paste0('EXCIT>', paste(x[!is.na(x)], collapse = ';'))}})

lin.anno.genes$gene.anno.inh <- apply(lin.anno.genes[,grepl(colnames(lin.anno.genes), pattern = 'Inhibitory')], 
                                  MARGIN = 1, FUN = function(x) {
    if (all(is.na(x))){NA}else{paste0('INHIB>', paste(x[!is.na(x)], collapse = ';'))}})

lin.anno.genes$gene.anno <- apply(lin.anno.genes[,c('gene.anno.exc','gene.anno.inh')], 
                                  MARGIN = 1, FUN = function(x) {
    if (all(is.na(x))){NA}else{paste(x[!is.na(x)], collapse = ' || ')}})

# apply(lin.anno.genes[,-(1:4)], MARGIN = 2, FUN = unique)


# Classify clusters by lineage annotations (Excitatory/Inhibitory)
lineages <- list(
    'Excitatory' = c(
        'NT-VGLUT1', 'NT-VGLUT2', 'NT-VGLUT3', # GLUT
        'P-PALL-M', 'P-PALL' # EMX1
    ),
    'Inhibitory' = c(
        'NT-GABA', # GAD2
        'P-DLGE', 'P-VLGE', 'P-SUBPALL', # DLX2
        'P-MGE-PO' # GAD2
    )
)

lin.annos <- st2[,c('ClusterID.(PoolClean)','NCells','AutoClass','Class','Subclass')]

for (lin in names(lineages)){
    for (ct in lineages[[lin]]){
        
        coincident <- grepl(pattern = ct, x = st2$AutoAnnotation,fixed = T)
        
        lin.annos <- cbind(lin.annos, ifelse(coincident, lin, NA))
        colnames(lin.annos)[ncol(lin.annos)] <- paste(lin, ct, sep='.')
        
    }
}

lin.annos$annot <- apply(lin.annos[,-(1:3)], 1, function(x){
    paste(unique(x[!is.na(x)]), collapse='::')
})

merge(lin.annos[,c(1,ncol(lin.annos))], lin.anno.genes[,-(2:4)], by = 'ClusterID.(PoolClean)') %>%
merge(y = st2, by = 'ClusterID.(PoolClean)') -> annotation.table

# Check self-consistency within lineage and gene annotations
annotation.table$annot.selfconsistent <- !(grepl('Inhibitory', x = annotation.table$annot, fixed=T)&grepl('Excitatory', x = annotation.table$annot, fixed=T))

annotation.table$geneannot.selfconsistent <- !(grepl('INHIB>', x = annotation.table$gene.anno, fixed=T)&grepl('EXCIT>', x = annotation.table$gene.anno, fixed=T))


# Check consistency between lineage and gene annotations
annotation.table$annot.betweenconsistent <- !(grepl('Inhibitory', x = annotation.table$annot, fixed=T)&grepl('EXCIT>', x = annotation.table$gene.anno, fixed=T)) & !(grepl('Excitatory', x = annotation.table$annot, fixed=T)&grepl('INHIB>', x = annotation.table$gene.anno, fixed=T))

# Define overall consistency based on self and between-column checks
annotation.table$consistent.all <- annotation.table$annot.betweenconsistent & annotation.table$annot.selfconsistent & annotation.table$geneannot.selfconsistent

all.cols <- unique(c(colnames(st2), colnames(lin.annos),colnames(lin.anno.genes)))
all.cols <- all.cols[all.cols %in% colnames(annotation.table)]
first.cols <- c('PoolOrder','ClusterID.(PoolClean)','NCells','AutoClass','Class','Subclass', 'consistent.all', 'annot', 'gene.anno', 'annot.selfconsistent', 'geneannot.selfconsistent', 'annot.betweenconsistent')
first.cols <- first.cols[first.cols%in%colnames(annotation.table)]
all.cols <- all.cols[!all.cols %in% first.cols]

annotation.table <- annotation.table[, c(first.cols, all.cols)]

openxlsx::write.xlsx(annotation.table, './cluster_lineage_annotation.xlsx')

# check inconsistencies
subset(annotation.table, !consistent.all)

# Part 2: Solve inconsistencies manually

# res: Dataframe holding the manually reviewed and commented annotations.
res <- openxlsx::read.xlsx('./cluster_lineage_annotation.commented.xlsx')
head(res)

res <- res[,c('ClusterID.(PoolClean)', 'Class', 'Subclass', 'Diagnotic', 'annot', 'gene.anno', 'consistent.all')]

# Combine 'Class' and 'Subclass' columns into 'ClassSubclass'
res$ClassSubclass <- apply(res[,c('Class', 'Subclass')], 1, function(x){paste(unique(x), collapse = ':')})

res$discarded <- (res$Diagnotic == 'DUDOSO') & !is.na(res$Diagnotic)

res$do.InEx <- res$Class %in% c('Neuroblast', 'Neuron', 'Neuronal IPC')
res$InEx <- NA

# Resolve inconsistent cases using 'Diagnotic' column
mask <- (!res$consistent.all)&(!res$discarded)&res$do.InEx
res[mask,'InEx'] <- res[mask,'Diagnotic']
res$InEx <- ifelse(is.na(res$InEx), NA, sapply(res$InEx, substr, start = 1, stop = 2))
# Assign In/Ex based on consistent 'annot' column
mask <- (res$consistent.all)&(!res$discarded)&res$do.InEx
res[mask,'InEx'] <- ifelse(grepl(res[mask,'annot'], pattern = 'Inhibitory', fixed=T), 'In', 'Ex')
# Assign remaining In/Ex based on 'gene.anno' column for consistent cases
mask <- (res$consistent.all)&(!res$discarded)&res$do.InEx&is.na(res$InEx)
res[mask,'InEx'] <- ifelse(grepl(res[mask,'gene.anno'], pattern = 'INHIB', fixed=T), 'In', 'Ex')

short.names <- c(
    'COPs (premyelinating)' = 'COPs',
    'Endothelial' = 'Endo',
    'Erythrocyte' = 'Eryt',
    'Fibroblast' = 'Fibr',
    'Glioblast' = 'GlioBl',
    'Immune' = 'Imm',
    'Neuroblast' = 'NeuBl',
    'Neuron' = 'Neu',
    'Neuronal IPC' = 'NeuIPC',
    'OPC' = 'OPC',
    'Pericytes' = 'Peri',
    'Placodes' = 'Plac',
    'Radial glia' = 'RGC',
    'Schwann' = 'Schwann',
    'VSMC' = 'VSMC')


table(res[,c('do.InEx','InEx')])
table(res[,c('Subclass','InEx')])


# Note: Using 'Subclass' for final naming convention
# Generate long name identifier (In/Ex + Class:Subclass)
res$LongName <- ifelse(
    res$discarded, NA, apply(
        res[,c('InEx', 'ClassSubclass')], 1, 
        function(x){
            paste(x[!is.na(x)], collapse = '')}))
    
# Generate short name identifier (In/Ex + Abbreviated Subclass)
res$ShortName <- ifelse(
    res$discarded, NA, apply(
        res[,c('InEx', 'Subclass')], 1, 
        function(x){
            x[2] <- short.names[x[2]]
            paste(x[!is.na(x)], collapse = '')}))


openxlsx::write.xlsx(res, './cluster_lineage_annotation.summary.xlsx')
