==> analyses/Braun_analyses/5_run_ewce.ipynb <==
# Set-up

# Define base directory for input data
base.data.dir <- "/users/genomics/xoel/Review_CD/ewce_linnarson.v7/data"


library(dplyr)
library(glue)

rbind.fill <- plyr::rbind.fill

my_time <- function(t0 = NULL){
    now <- Sys.time()
    if(!is.null(t0)){
        diff <- now -t0
        message(glue('[{now}] (+{diff}s)'))
    } else {
        message(glue('[{now}]'))
    }
    return()
}

# **Part 1: List the CTDs to analyse**

# Define input directory for CTD files
in.dir <- file.path(base.data.dir, "by_Subregion")

ctd.files <- list.files(in.dir, full.names = T, recursive = T) %>% grep(value = T, pattern = 'ctd_', fixed=T) %>% sort
ctd.files
length(ctd.files)

# **Part 2: List the lists to analyse**

# results_name <-paste0('Diseases_',Sys.Date())
results_name <-'Diseases_2024-11-28'
results_name

# Define path to gene list file
gene.list.file <- '/users/genomics/xoel/codebases/gene_disease_asociation/parsed_lists_with_nicola.union.csv'

disease.associations <- data.table::fread(
    file = gene.list.file, 
    data.table=F) %>% rename(Gene=V1)
disease.associations.lists <- lapply(colnames(disease.associations)[-1], 
                                     function(x){disease.associations$Gene[as.logical(disease.associations[,x])]}) %>% 
    setNames(nm=colnames(disease.associations)[-1])
diseases <- names(disease.associations.lists)

# **Part 3: Run EWCE**

reps <- 10000
overwrite <- F

lists <-  disease.associations.lists
list.names <- names(lists)

for (ctd.file in ctd.files){

    path <- dirname(ctd.file)
    t0 <- Sys.time()
    
    my_time(t0)

    res.file <- glue('{path}/EWCE.{results_name}.{reps}_reps.results.csv')
    print(res.file)    
    
    if (overwrite | !file.exists(res.file)){
        message('Running because results were not found or overwrite is set to T')

        load(ctd.file, verbose=T)
                
        n.levels <- length(ctd)
        print(glue('There are {n.levels} annotation levels in this CTD'))
        
        #########################################################
        # FOR N LEVELS
        # Bootstrap significance test, no control for transcript length and GC content
        lapply(1:n.levels, function(annotLevel){
        
            #########################################################
            # FOR EACH LIST
            level.results <- lapply(1:length(list.names), function(list.i){
                list.name <- list.names[list.i]
                print(glue('{list.i+(annotLevel-1)*length(list.names)}/{length(list.names)*n.levels}:\t{list.name}'))
                my_time(t0)
                
                full_results <- EWCE::bootstrap_enrichment_test(sct_data = ctd, 
                                                                geneSizeControl = F,
                                                                no_cores = parallel::detectCores(),
                                                                sctSpecies = "human",
                                                                genelistSpecies = "human",
                                                                hits = lists[[list.name]], 
                                                                reps = reps,
                                                                annotLevel = annotLevel)
                for (a in names(full_results)){
                    if (!is.data.frame(full_results[[a]])){
                        full_results[[a]] <- data.frame(hit.cells = full_results[[a]]) %>% t() %>% as.data.frame()
                    }
                    full_results[[a]] <- full_results[[a]] %>% mutate(
                    annotLevel = annotLevel, list = list.name)
                }
                
                return(full_results)
                
            }) %>% setNames(nm = list.names)
            #########################################################
            gc()
            # Aggregate results across lists for the current annotation level
            lapply(names(level.results[[1]]), function(x){
                lapply(level.results, function(y){y[[x]]}) %>% do.call(what='rbind')
            }) %>% setNames(names(level.results[[1]])) -> level.results
            
            return(level.results)
            
        }) -> full_results
        #########################################################
        my_time(t0)
        
        # Aggregate results across all annotation levels
        lapply(names(full_results[[1]]), function(x){
            lapply(full_results, function(y){y[[x]]}) %>% do.call(what='rbind.fill')
        }) %>% setNames(names(full_results[[1]])) -> full_results
            
        #########################################################
        # Save
        for (iname in names(full_results)){
            data.table::fwrite(full_results[[iname]], glue('{path}/EWCE.{results_name}.{reps}_reps.{iname}.csv'))
        }
        
        ctd <- NULL
        full_results <- NULL
        
    } else  {
        message('Not re-running because results were found or overwrite is set to F')
    }
}
